+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * `Function.enforceProperty` will already add the used symbol to the prototype
 * Make `String.decodeAttributes()` decode tag attributes by default (no separator)
 * Fix `Date#timeAgo()` returning big hours, minutes & seconds when date is more than a month ago
+* Make `Array#sortByPath()` handle comparing `undefined` and objects better
 
 ## 0.6.6 (2019-02-25)